requirements_torch+tensor.txt
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.61.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-python==4.9.0.80
packaging==25.0
pandas==2.3.3
pillow==11.3.0
polars==1.35.2
polars-runtime-32==1.35.2
psutil==7.1.3
pycocotools==2.0.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.3
seaborn==0.13.2
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorflow==2.15.1
thop==0.1.1.post2209072238
# install torch stack via the PyTorch CUDA 11.8 wheel index: https://download.pytorch.org/whl/cu118
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
triton==2.2.0
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.235
ultralytics-thop==2.0.18
urllib3==2.5.0
